Adding defc to the external data types
+ Also only loading DEFC when there are changes

+++ dataactvalidator/scripts/load_defc.py
@@ -1,10 +1,16 @@
 import requests
+import sys
 import os
 import logging
 import pandas as pd
 
+from datetime import datetime
+
+from dataactbroker.helpers.pandas_helper import check_dataframe_diff
+
 from dataactcore.logging import configure_logging
 from dataactcore.interfaces.db import GlobalDB
+from dataactcore.interfaces.function_bag import update_external_data_load_date
 from dataactcore.config import CONFIG_BROKER
 from dataactcore.models.domainModels import DEFC
 
@@ -14,9 +20,15 @@ from dataactvalidator.scripts.loader_utils import clean_data, insert_dataframe
 logger = logging.getLogger(__name__)
 
 
-def load_defc():
-    """ Loads the DEFC data. """
+def load_defc(force_reload=False):
+    """ Loads the DEFC data.
+
+        Args:
+            force_reload: boolean to determine if reload should happen whether there are differences or not
+    """
+    start_time = datetime.now()
     defc_file = os.path.join(CONFIG_BROKER['path'], 'dataactvalidator', 'config', 'def_codes.csv')
+
     try:
         # Update file from public S3 bucket
         def_codes_url = '{}/def_codes.csv'.format(CONFIG_BROKER['usas_public_reference_url'])
@@ -39,16 +51,23 @@ def load_defc():
             {}
         )
 
-        sess = GlobalDB.db().session
-        # delete any data in the DEFC table
-        sess.query(DEFC).delete()
+        diff_found = check_dataframe_diff(data, DEFC, ['defc_id'], ['code'])
 
-        # insert data into table
-        num = insert_dataframe(data, DEFC.__table__.name, sess.connection())
-        logger.info('{} records inserted to defc'.format(num))
-        sess.commit()
+        if force_reload or diff_found:
+            sess = GlobalDB.db().session
+            # delete any data in the DEFC table
+            sess.query(DEFC).delete()
+
+            # insert data into table
+            num = insert_dataframe(data, DEFC.__table__.name, sess.connection())
+            logger.info('{} records inserted to defc'.format(num))
+            sess.commit()
+            update_external_data_load_date(start_time, datetime.now(), 'defc')
+        else:
+            logger.info('No differences found, skipping defc table reload.')
 
 
 if __name__ == '__main__':
     configure_logging()
-    load_defc()
+    reload = '--force' in sys.argv
+    load_defc(reload)
